requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio 
appnope 
argon2-cffi 
argon2-cffi-bindings 
asttokens 
async-lru 
attrs 
babel 
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4 
bleach 
brotlicffi 
build==1.3.0
cachetools==5.5.2
certifi 
cffi 
charset-normalizer 
chromadb==1.0.17
click==8.2.1
coloredlogs==15.0.1
comm 
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.0.0
debugpy 
decorator 
defusedxml 
dill==0.3.8
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
emoji==2.14.1
executing 
faiss==1.9.0
fastjsonschema 
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.3.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11 
hf-xet==1.1.7
html5lib==1.1
httpcore 
httptools==0.6.4
httpx 
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna 
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel 
ipython 
ipython_pygments_lexers 
jedi 
Jinja2 
jiter==0.10.0
joblib==1.5.1
json5 
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema 
jsonschema-specifications 
jupyter-events 
jupyter-lsp 
jupyter_client 
jupyter_core 
jupyter_server 
jupyter_server_terminals 
jupyterlab 
jupyterlab_pygments 
jupyterlab_server 
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langdetect==1.0.9
langgraph==0.6.5
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.14
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe 
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline 
mdurl==0.1.2
mistune 
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient 
nbconvert 
nbformat 
nest-asyncio 
networkx==3.5
nltk==3.9.1
notebook_shim 
numpy==2.3.2
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.1
openai==1.99.9
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
ormsgpack==1.10.0
overrides 
packaging 
pandas==2.3.1
pandocfilters 
parso 
pdfminer.six==20250506
pexpect 
pillow==11.3.0
platformdirs 
posthog==5.4.0
primp==0.15.0
prometheus_client 
prompt-toolkit 
propcache==0.3.2
protobuf==6.32.0
psutil 
ptyprocess 
pure-eval 
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser 
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments 
pyparsing==3.2.3
pypdf==6.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks 
python-dateutil 
python-dotenv==1.1.1
python-iso639==2025.2.18
python-json-logger 
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML 
pyzmq 
RapidFuzz==3.13.0
referencing 
regex==2025.7.34
requests 
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator 
rfc3986-validator 
rich==14.1.0
rpds-py 
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
Send2Trash 
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
six 
sniffio 
soupsieve 
SQLAlchemy==2.0.43
stack-data 
sympy==1.14.0
tenacity==9.1.2
terminado 
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2 
tokenizers==0.21.4
torch==2.8.0
tornado 
tqdm==4.67.1
traitlets 
transformers==4.55.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions 
tzdata==2025.2
unstructured==0.18.13
unstructured-client==0.42.3
urllib3 
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth 
webencodings 
websocket-client 
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
